+++ alicloud/diff_suppress_funcs.go
@@ -1,6 +1,7 @@
 package alicloud
 
 import (
+	"github.com/denverdino/aliyungo/common"
 	"github.com/denverdino/aliyungo/dns"
 	"github.com/denverdino/aliyungo/slb"
 	"github.com/hashicorp/terraform/helper/schema"
@@ -81,4 +82,56 @@ func dnsPriorityDiffSuppressFunc(k, old, new string, d *schema.ResourceData) boo
 		return false
 	}
 	return true
+}
+
+func slbInternetDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	if internet, ok := d.GetOk("internet"); ok && internet.(bool) {
+		return true
+	}
+	return false
+}
+
+func slbInternetChargeTypeDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	return !slbInternetDiffSuppressFunc(k, old, new, d)
+}
+
+func slbBandwidthDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	if slbInternetDiffSuppressFunc(k, old, new, d) && slb.InternetChargeType(d.Get("internet_charge_type").(string)) == slb.PayByBandwidth {
+		return false
+	}
+	return true
+}
+
+func ecsPrivateIpDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	vswitch := ""
+	if vsw, ok := d.GetOk("vswitch_id"); ok && vsw.(string) != "" {
+		vswitch = vsw.(string)
+	} else if subnet, ok := d.GetOk("subnet_id"); ok && subnet.(string) != "" {
+		vswitch = subnet.(string)
+	}
+
+	if vswitch != "" {
+		return false
+	}
+	return true
+}
+func ecsInternetDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	if allocate, ok := d.GetOk("allocate_public_ip"); ok && allocate.(bool) {
+		return false
+	}
+	return true
+}
+
+func ecsPostPaidDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	if common.InstanceChargeType(d.Get("instance_charge_type").(string)) == common.PrePaid {
+		return false
+	}
+	return true
+}
+
+func ecsChargeTypeSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	if common.InstanceChargeType(old) == common.PrePaid && common.InstanceChargeType(new) == common.PostPaid {
+		return true
+	}
+	return false
 }